Add files via upload
I created an add comment/deduct points button (the first big text box is the comment, the second little one is the points reduction). I know it looks ugly but it's all I could do with the 1hr/week commitment. It will look better in the future. Lastly, I added a way to save the comments as a text file. I will change it to be able to do a PDF instead but I wanted to see if I could even do the text file first.